### 2023-09-14 #5847 @editorialbot
[PRE REVIEW]: Qiskit Dynamics: A Python package for simulating the time dynamics of quantum systems
**Submitting author:** @DanPuzzuoli (<a href="http://orcid.org/0000-0002-3815-2072">Daniel Puzzuoli</a>)
**Repository:** https://github.com/Qiskit-Extensions/qiskit-dynamics
**Branch with paper.md** (empty if default branch): paper
**Version:** 0.4.2
