### 2023-09-14 #5847 @danielskatz
👋 @DanPuzzuoli - I also see that your references didn't build correctly.  Please see the [example paper](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography). You've quoted all the references, e.g.,
```
`[@alexander_qiskit_2020]`
```
instead of just using brackets without quotes, e.g.
```
[@alexander_qiskit_2020]
```

you can change these and then recompile with `@editorialbot generate pdf` as mentioned above
