### 2023-09-14 79715ef
Fix citation format and missing DOI (#258)
