+++ paper.bib
@@ -14,10 +14,11 @@
 }
 
 @inproceedings{torchquantum,
-    title     = {Quantumnas: Noise-adaptive search for robust quantum circuits},
-    author    = {Wang, Hanrui and Ding, Yongshan and Gu, Jiaqi and Li, Zirui and Lin, Yujun and Pan, David Z and Chong, Frederic T and Han, Song},
-    booktitle = {The 28th IEEE International Symposium on High-Performance Computer Architecture (HPCA-28)},
-    year      = {2022}
+    title={Quantumnas: Noise-adaptive search for robust quantum circuits},
+    author={Wang, Hanrui and Ding, Yongshan and Gu, Jiaqi and Li, Zirui and Lin, Yujun and Pan, David Z and Chong, Frederic T and Han, Song},
+    booktitle={The 28th IEEE International Symposium on High-Performance Computer Architecture (HPCA-28)},
+    year={2022},
+	doi={10.1109/hpca53966.2022.00057}
 }
 
 @article{C3,
