### 2023-09-14 #5847 @DanPuzzuoli
Hi @danielskatz, thank you for being the editor on this submission! Also thanks to @babreu-ncsa for being a reviewer.

I believe I've fixed the citation issues (both the missing DOI and the erroneous quotation marks - I'm not sure why I thought the quotes were necessary).

For reviewer suggestions:

- goerz: they are in the list of pre-existing JOSS reviewers, and are coincidentally the author of one of the packages cited in the paper, so they would be a good fit.
- hodgestar or nwlambert: Both are major contributors to QuTiP, which is the largest package offering similar functionality to this submission.
- nwittler: A major contributor to C3 (another related package cited in the paper).
