### 2023-09-14 #5847 @danielskatz
👋 @goerz - would you be willing to review this JOSS submission, given your connection to the topic/software community?
