### 2023-09-14 #5847 @danielskatz
👋 @hodgestar and @nwlambert - would one of you be willing to review this JOSS submission?  We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
