### 2023-09-29 #5853 @danielskatz
@babreu-ncsa - can you generate your checklist too, and at least check off the conflict of interest and code of conduct items, to make sure things are working?
