### 2023-09-30 d39b96b
add global numpy/scipy aliases
