+++ test/dynamics/arraylias/test_alias.py
@@ -9,7 +9,7 @@
 # Any modifications or derivative works of this code must retain this
 # copyright notice, and modified files need to carry a notice indicating
 # that they have been altered from the originals.
-# pylint: disable=invalid-name
+# pylint: disable=invalid-name,no-member
 
 """
 Test global alias instances.
@@ -17,13 +17,13 @@ Test global alias instances.
 
 from functools import partial
 
-from ..common import QiskitDynamicsTestCase, test_array_backends
-
 import numpy as np
 import scipy as sp
 
 from qiskit_dynamics import DYNAMICS_NUMPY as unp
 from qiskit_dynamics import DYNAMICS_SCIPY as usp
+
+from ..common import QiskitDynamicsTestCase, test_array_backends
 
 
 @partial(test_array_backends, backends=["numpy", "jax", "array_numpy", "array_jax"])
@@ -32,11 +32,11 @@ class TestDynamicsNumpy(QiskitDynamicsTestCase):
 
     def test_simple_case(self):
         """Validate correct type and output."""
-        a = self.asarray([1., 2., 3.])
+        a = self.asarray([1.0, 2.0, 3.0])
         output = unp.exp(a)
-        self.assertTrue(isinstance(output, type(a)))
+        self.assertArrayType(output)
 
-        expected = np.exp(np.array([1., 2., 3.]))
+        expected = np.exp(np.array([1.0, 2.0, 3.0]))
         self.assertAllClose(output, expected)
 
 
@@ -46,9 +46,9 @@ class TestDynamicsScipy(QiskitDynamicsTestCase):
 
     def test_simple_case(self):
         """Validate correct type and output."""
-        a = self.asarray([1., 2., 3.])
+        a = self.asarray([1.0, 2.0, 3.0])
         output = usp.fft.dct(a)
-        self.assertTrue(isinstance(output, type(a)))
+        self.assertArrayType(output)
 
-        expected = sp.fft.dct(np.array([1., 2., 3.]))
+        expected = sp.fft.dct(np.array([1.0, 2.0, 3.0]))
         self.assertAllClose(output, expected)
