### 2023-10-01 4d3f6c2
removing unnecessary argument
