### 2023-10-01 159a22b
removing unnecessary partial call
