+++ test/dynamics/arraylias/test_alias.py
@@ -40,7 +40,7 @@ class TestDynamicsNumpy(QiskitDynamicsTestCase):
         self.assertAllClose(output, expected)
 
 
-@partial(test_array_backends)
+@test_array_backends
 class TestDynamicsScipy(QiskitDynamicsTestCase):
     """Test cases for global scipy configuration."""
 
