### 2023-10-03 d553fbf
Fix citation of JuliaQuantumControl
