+++ paper.bib
@@ -173,7 +173,8 @@
 }
 
 @misc{julia_qc,
-  title = {A Julia Framework for Quantum Optimal Control},
+  author = {Goerz, Michael H. and Carrasco, Sebastián C. and Malinovsky, Vladimir S. and {Contributors}},
+  title = {A {Julia} Framework for Quantum Optimal Control},
   year = {2023},
   publisher = {GitHub},
   journal = {GitHub repository},
