### 2023-10-03 #5853 @goerz
The paper and underlying software package are of exemplary quality. I have no objections to publication, up to two trivial "typo" issues in the manuscript (which is why I've left the "Quality of writing" and "References" unchecked for now).

https://github.com/Qiskit-Extensions/qiskit-dynamics/pull/261
https://github.com/Qiskit-Extensions/qiskit-dynamics/pull/262
