### 2023-10-03 444c517
Fix typo "v.s." (should be "vs.") (#262)
The abbreviation "v.s." stands for "vidre supra" ("see above"), not
"versus", which should be abbreviated as "vs." in non-legal contexts (or
"vs", depending on the style guide, but I think "vs." is far more common
in American English)

As a smaller issue, "e.g." is always followed by a comma.
