+++ paper.md
@@ -37,7 +37,7 @@ bibliography: paper.bib
 
 # Summary
 
-Qiskit Dynamics is an open source Python library for numerically simulating the time dynamics of finite-dimensional quantum systems. The goal of the package is to provide flexible configuration of the numerical methods used for simulation: general tools for transforming models of quantum systems for more efficient simulation (rotating frames and the rotating wave approximation), choice of array representations (dense v.s. sparse, and different array libraries), and access to different types of underlying solvers (standard ODE v.s. geometric solvers). The package also contains advanced functionality for computing time-dependent perturbation theory expressions used in robust quantum control optimization [@perturb1; @perturb2]. 
+Qiskit Dynamics is an open source Python library for numerically simulating the time dynamics of finite-dimensional quantum systems. The goal of the package is to provide flexible configuration of the numerical methods used for simulation: general tools for transforming models of quantum systems for more efficient simulation (rotating frames and the rotating wave approximation), choice of array representations (dense vs. sparse, and different array libraries), and access to different types of underlying solvers (standard ODE vs. geometric solvers). The package also contains advanced functionality for computing time-dependent perturbation theory expressions used in robust quantum control optimization [@perturb1; @perturb2].
 
 As part of the Qiskit Ecosystem (https://qiskit.org/ecosystem), the package interfaces with other parts of Qiskit [@Qiskit]. Most notably, Qiskit Dynamics provides tools for simulating control sequences specified by Qiskit Pulse [@alexander_qiskit_2020] which gives a pulse representation of quantum circuit instructions. Higher level interfaces allow users to build and interact with simulation-based objects that target the same constraints (qubit layout, control sequence sampling rate, etc.) as a specified IBM Quantum computer.
 
@@ -45,7 +45,7 @@ Lastly, to facilitate high-perfomance applications, Qiskit Dynamics is compatibl
 
 # Statement of need
 
-Numerical simulation of time-dependent quantum systems is a useful tool in quantum device characterization, design, and control optimization. As these applications often involve the expensive process of repeatedly simulating a system across different parameters (e.g. in exploratory parameter scans or in optimizations), users need to be able to easily select the numerical methods that are most performant for their specific problem. The ability to automatically differentiate and compile simulations is also critical for control optimization research. To ensure flexibility and broad applicability, it is important that all of these capabilities work for arbitrary user-specified models.
+Numerical simulation of time-dependent quantum systems is a useful tool in quantum device characterization, design, and control optimization. As these applications often involve the expensive process of repeatedly simulating a system across different parameters (e.g., in exploratory parameter scans or in optimizations), users need to be able to easily select the numerical methods that are most performant for their specific problem. The ability to automatically differentiate and compile simulations is also critical for control optimization research. To ensure flexibility and broad applicability, it is important that all of these capabilities work for arbitrary user-specified models.
 
 Furthermore, having a simulation-based drop-in replacement for real quantum computing systems is useful for developers building software tools for low-level control of experiments, such as Qiskit Pulse [@alexander_qiskit_2020] and Qiskit Experiments [@kanazawa_qiskit_2023].
 
