### 2023-10-06 db07e1b
Update test/dynamics/common.py
Co-authored-by: Kento Ueda <38037695+to24toro@users.noreply.github.com>
