### 2023-10-06 9efa7bd
adding more documentation explaining the testing infrastructure
