### 2023-10-08 e0a48d0
Update test/dynamics/common.py
Co-authored-by: Kento Ueda <38037695+to24toro@users.noreply.github.com>
