### 2023-10-11 842d735
Add global instances of NumPy/SciPy aliases (#260)
Co-authored-by: Kento Ueda <38037695+to24toro@users.noreply.github.com>
