+++ test/dynamics/common.py
@@ -12,17 +12,40 @@
 # pylint: disable=invalid-name,isinstance-second-argument-not-valid-type
 
 """
-Shared functionality and helpers for the unit tests.
+General infrastructure for unit tests.
+
+``QiskitDynamicsTestCase`` adds general functionality used across tests.
+
+The ``test_array_backends`` decorator, along with the classes with names ``<array_library>TestBase``
+enable writing test cases that are agnostic to the underlying array library. The general setup is:
+- Each ``<array_library>TestBase`` implements an ``array_library`` class method returning a string,
+  an ``asarray`` method for specifically defining arrays for that library, and an
+  ``assertArrayType`` method for validating that an array is from that library. These classes can
+  also implement any required setup and teardown methods for working with that library (e.g.
+  ``JAXTestBase`` skips tests if running on windows).
+- When used on a given ``test_class``, the decorator ``test_array_backends`` creates a series of
+  subclasses inheriting from ``test_class`` and a desired list of ``<array_library>TestBase``. The
+  desired list is specified via the ``array_libraries`` argument, which are matched against the
+  output of ``<array_library>TestBase.array_library()``. Note that the actual format of the class
+  name ``<array_library>TestBase`` is irrelevant: ``test_array_backends`` automatically detects
+  classes in this file with ``array_library`` class methods.
+Using the above infrastructure, a new array library can be added to the testing infrastructure by
+adding a new ``<array_library>TestBase`` class implementing the required methods, and possibly
+adding the output of ``<array_library>TestBase.array_library()`` to the default list of array
+libraries in ``test_array_backends``.
 """
 
+from typing import Type, Optional, List, Callable, Iterable
 import warnings
 import unittest
-from typing import Callable, Iterable
+import inspect
+
 import numpy as np
 from scipy.sparse import issparse
 
 try:
     from jax import jit, grad
+    import jax.numpy as jnp
 except ImportError:
     pass
 
@@ -34,8 +57,8 @@ class QiskitDynamicsTestCase(unittest.TestCase):
 
     def assertAllClose(self, A, B, rtol=1e-8, atol=1e-8):
         """Call np.allclose and assert true."""
-        A = Array(A)
-        B = Array(B)
+        A = np.array(A)
+        B = np.array(B)
 
         self.assertTrue(np.allclose(A, B, rtol=rtol, atol=atol))
 
@@ -53,6 +76,170 @@ class QiskitDynamicsTestCase(unittest.TestCase):
         self.assertTrue(np.allclose(A, B, rtol=rtol, atol=atol))
 
 
+class NumpyTestBase(unittest.TestCase):
+    """Base class for tests working with numpy arrays."""
+
+    @classmethod
+    def array_library(cls):
+        """Library method."""
+        return "numpy"
+
+    def asarray(self, a):
+        """Array generation method."""
+        return np.array(a)
+
+    def assertArrayType(self, a):
+        """Assert the correct array type."""
+        return isinstance(a, np.ndarray)
+
+
+class JAXTestBase(unittest.TestCase):
+    """Base class for tests working with JAX arrays."""
+
+    @classmethod
+    def setUpClass(cls):
+        try:
+            # pylint: disable=import-outside-toplevel
+            import jax
+
+            jax.config.update("jax_enable_x64", True)
+            jax.config.update("jax_platform_name", "cpu")
+        except Exception as err:
+            raise unittest.SkipTest("Skipping jax tests.") from err
+
+    @classmethod
+    def array_library(cls):
+        """Library method."""
+        return "jax"
+
+    def asarray(self, a):
+        """Array generation method."""
+        return jnp.array(a)
+
+    def assertArrayType(self, a):
+        """Assert the correct array type."""
[…]
