### 2023-10-11 b3c4471
modifying tests to directly test time values of variable step solvers
