+++ test/dynamics/solvers/test_solver_classes.py
@@ -296,12 +296,12 @@ class TestSolverSignalHandling(QiskitDynamicsTestCase):
         t_span = [0.0, 1.42]
         signals = [Signal(3.0)]
 
-        res1 = ham_solver.solve(t_span=t_span, y0=y0, signals=signals)
+        res1 = ham_solver.solve(t_span=t_span, y0=y0, signals=signals, atol=1e-10, rtol=1e-10)
 
         self.ham_model.signals = signals
-        res2 = solve_lmde(generator=self.ham_model, t_span=t_span, y0=y0)
+        res2 = solve_lmde(generator=self.ham_model, t_span=t_span, y0=y0, atol=1e-10, rtol=1 - 10)
 
-        self.assertAllClose(res1.y, res2.y)
+        self.assertAllClose(res1.y[-1], res2.y[-1], atol=1e-8, rtol=1e-8)
 
     def test_lindblad_model(self):
         """Test for solver with only static lindblad information."""
@@ -317,12 +317,14 @@ class TestSolverSignalHandling(QiskitDynamicsTestCase):
         t_span = [0.0, 1.42]
         signals = [Signal(3.0)]
 
-        res1 = lindblad_solver.solve(t_span=t_span, y0=y0, signals=signals)
+        res1 = lindblad_solver.solve(t_span=t_span, y0=y0, signals=signals, atol=1e-10, rtol=1e-10)
 
         self.lindblad_model.signals = (signals, None)
-        res2 = solve_lmde(generator=self.lindblad_model, t_span=t_span, y0=y0)
+        res2 = solve_lmde(
+            generator=self.lindblad_model, t_span=t_span, y0=y0, atol=1e-10, rtol=1e-10
+        )
 
-        self.assertAllClose(res1.y, res2.y)
+        self.assertAllClose(res1.y[-1], res2.y[-1], atol=1e-8, rtol=1e-8)
 
     def test_td_lindblad_model(self):
         """Test for solver with time-dependent lindblad information."""
@@ -339,12 +341,16 @@ class TestSolverSignalHandling(QiskitDynamicsTestCase):
         t_span = [0.0, 1.42]
         signals = ([Signal(3.0)], [Signal(2.0)])
 
-        res1 = td_lindblad_solver.solve(t_span=t_span, y0=y0, signals=signals)
+        res1 = td_lindblad_solver.solve(
+            t_span=t_span, y0=y0, signals=signals, atol=1e-10, rtol=1e-10
+        )
 
         self.td_lindblad_model.signals = signals
-        res2 = solve_lmde(generator=self.td_lindblad_model, t_span=t_span, y0=y0)
+        res2 = solve_lmde(
+            generator=self.td_lindblad_model, t_span=t_span, y0=y0, atol=1e-10, rtol=1e-10
+        )
 
-        self.assertAllClose(res1.y, res2.y)
+        self.assertAllClose(res1.y[-1], res2.y[-1], atol=1e-8, rtol=1e-8)
 
     def test_rwa_ham_model(self):
         """Test correct handling of RWA for a Hamiltonian model."""
@@ -361,13 +367,13 @@ class TestSolverSignalHandling(QiskitDynamicsTestCase):
         t_span = [0.0, 1.0]
         signals = [Signal(1.0, carrier_freq=5.0)]
 
-        res1 = rwa_ham_solver.solve(t_span=t_span, y0=y0, signals=signals)
+        res1 = rwa_ham_solver.solve(t_span=t_span, y0=y0, signals=signals, atol=1e-10, rtol=1e-10)
 
         self.ham_model.signals = signals
         rwa_ham_model = rotating_wave_approximation(self.ham_model, cutoff_freq=5.0)
-        res2 = solve_lmde(generator=rwa_ham_model, t_span=t_span, y0=y0)
[…]
