+++ test/dynamics/solvers/test_solver_classes.py
@@ -296,10 +296,10 @@ class TestSolverSignalHandling(QiskitDynamicsTestCase):
         t_span = [0.0, 1.42]
         signals = [Signal(3.0)]
 
-        res1 = ham_solver.solve(t_span=t_span, y0=y0, signals=signals, atol=1e-10, rtol=1e-10)
+        res1 = ham_solver.solve(t_span=t_span, y0=y0, signals=signals, atol=1e-11, rtol=1e-11)
 
         self.ham_model.signals = signals
-        res2 = solve_lmde(generator=self.ham_model, t_span=t_span, y0=y0, atol=1e-10, rtol=1 - 10)
+        res2 = solve_lmde(generator=self.ham_model, t_span=t_span, y0=y0, atol=1e-11, rtol=1e-11)
 
         self.assertAllClose(res1.y[-1], res2.y[-1], atol=1e-8, rtol=1e-8)
 
@@ -317,11 +317,11 @@ class TestSolverSignalHandling(QiskitDynamicsTestCase):
         t_span = [0.0, 1.42]
         signals = [Signal(3.0)]
 
-        res1 = lindblad_solver.solve(t_span=t_span, y0=y0, signals=signals, atol=1e-10, rtol=1e-10)
+        res1 = lindblad_solver.solve(t_span=t_span, y0=y0, signals=signals, atol=1e-11, rtol=1e-11)
 
         self.lindblad_model.signals = (signals, None)
         res2 = solve_lmde(
-            generator=self.lindblad_model, t_span=t_span, y0=y0, atol=1e-10, rtol=1e-10
+            generator=self.lindblad_model, t_span=t_span, y0=y0, atol=1e-11, rtol=1e-11
         )
 
         self.assertAllClose(res1.y[-1], res2.y[-1], atol=1e-8, rtol=1e-8)
@@ -342,12 +342,12 @@ class TestSolverSignalHandling(QiskitDynamicsTestCase):
         signals = ([Signal(3.0)], [Signal(2.0)])
 
         res1 = td_lindblad_solver.solve(
-            t_span=t_span, y0=y0, signals=signals, atol=1e-10, rtol=1e-10
+            t_span=t_span, y0=y0, signals=signals, atol=1e-11, rtol=1e-11
         )
 
         self.td_lindblad_model.signals = signals
         res2 = solve_lmde(
-            generator=self.td_lindblad_model, t_span=t_span, y0=y0, atol=1e-10, rtol=1e-10
+            generator=self.td_lindblad_model, t_span=t_span, y0=y0, atol=1e-11, rtol=1e-11
         )
 
         self.assertAllClose(res1.y[-1], res2.y[-1], atol=1e-8, rtol=1e-8)
@@ -367,11 +367,11 @@ class TestSolverSignalHandling(QiskitDynamicsTestCase):
         t_span = [0.0, 1.0]
         signals = [Signal(1.0, carrier_freq=5.0)]
 
-        res1 = rwa_ham_solver.solve(t_span=t_span, y0=y0, signals=signals, atol=1e-10, rtol=1e-10)
+        res1 = rwa_ham_solver.solve(t_span=t_span, y0=y0, signals=signals, atol=1e-11, rtol=1e-11)
 
         self.ham_model.signals = signals
         rwa_ham_model = rotating_wave_approximation(self.ham_model, cutoff_freq=5.0)
-        res2 = solve_lmde(generator=rwa_ham_model, t_span=t_span, y0=y0, atol=1e-10, rtol=1e-10)
+        res2 = solve_lmde(generator=rwa_ham_model, t_span=t_span, y0=y0, atol=1e-11, rtol=1e-11)
 
[…]
