### 2023-10-12 ae2b009
Adding a missing "annihilation" to backend string parser doc (#265)
