### 2023-10-12 03a48e0
modifying test to only check final state
