### 2023-10-12 690c7c9
removing restrictions on JAX/Diffrax versions in tox/setup
