+++ tox.ini
@@ -15,18 +15,16 @@ commands = stestr run {posargs}
 [testenv:jax]
 deps =
     -r{toxinidir}/requirements-dev.txt
-    jax<=0.4.6
-    jaxlib<=0.4.6
-    equinox<=0.10.3
-    diffrax<=0.3.1
+    jax
+    jaxlib
+    diffrax
 
 [testenv:lint]
 deps =
     -r{toxinidir}/requirements-dev.txt
-    jax<=0.4.6
-    jaxlib<=0.4.6
-    equinox<=0.10.3
-    diffrax<=0.3.1
+    jax
+    jaxlib
+    diffrax
 commands =
   black --check {posargs} qiskit_dynamics test
   pylint -rn -j 0 --rcfile={toxinidir}/.pylintrc qiskit_dynamics/ test/
@@ -41,8 +39,8 @@ commands = black {posargs} qiskit_dynamics test
 usedevelop = False
 deps =
     -r{toxinidir}/requirements-dev.txt
-    jax<=0.4.6
-    jaxlib<=0.4.6
+    jax
+    jaxlib
     diffrax
 commands =
   sphinx-build -j auto -b html -W {posargs} docs/ docs/_build/html
