### 2023-10-13 #5853 @babreu-ncsa
There are some issues with installing the package along with JAX for OSX + Apple Silicon that may be worth a note in the [Installation](https://github.com/Qiskit-Extensions/qiskit-dynamics/tree/stable/0.4#installation) section of the repo's README.

```
$ conda create --name qiskit-dynamics-jax-cpu -c conda-forge python=3.11
$ conda activate qiskit-dynamics-jax-cpu
$ pip install "qiskit-dynamics[jax]"
$ python
>>> import jax
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "/Users/babreu/opt/anaconda3/envs/qiskit-dynamics-jax-cpu/lib/python3.11/site-packages/jax/__init__.py", line 35, in <module>
    from jax import config as _config_module
  File "/Users/babreu/opt/anaconda3/envs/qiskit-dynamics-jax-cpu/lib/python3.11/site-packages/jax/config.py", line 17, in <module>
    from jax._src.config import config  # noqa: F401
    ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/Users/babreu/opt/anaconda3/envs/qiskit-dynamics-jax-cpu/lib/python3.11/site-packages/jax/_src/config.py", line 25, in <module>
    from jax._src import lib
  File "/Users/babreu/opt/anaconda3/envs/qiskit-dynamics-jax-cpu/lib/python3.11/site-packages/jax/_src/lib/__init__.py", line 85, in <module>
    cpu_feature_guard.check_cpu_features()
RuntimeError: This version of jaxlib was built using AVX instructions, which your CPU and/or operating system do not support. You may be able work around this issue by building jaxlib from source.
```

```
$ system_profiler SPSoftwareDataType
Software:

    System Software Overview:

      System Version: macOS 13.6 (22G120)
      Kernel Version: Darwin 22.6.0
```

```
$ system_profiler SPHardwareDataType
Hardware:

    Hardware Overview:

      Model Name: MacBook Pro
      Model Identifier: MacBookPro18,2
      Model Number: Z14W0010BLL/A
      Chip: Apple M1 Max
```
