### 2023-10-13 #5853 @babreu-ncsa
There is just one thing blocking me from completing the review (everything else is great and I'm able to reproduce what's in the tutorials:

***Automated tests**: Are there automated tests or manual steps described so that the functionality of the software can be verified?*

I see there are tests in the repository (https://github.com/Qiskit-Extensions/qiskit-dynamics/tree/main/test), but it's not clear to me when these are run and if they can be triggered manually by the user.
