### 2023-10-13 #5853 @goerz
@babreu-ncsa They do run automatically via the [`main.yml` workflow](https://github.com/Qiskit-Extensions/qiskit-dynamics/blob/main/.github/workflows/main.yml).

You can run them locally using `tox` (running more or less the same commands as in the `main.yml`). I've done this on my laptop, and the tests pass.
