### 2023-10-16 02edfce
arraylias to signal class
