+++ test/dynamics/signals/test_signals.py
@@ -15,13 +15,15 @@
 Tests for signals.
 """
 
+from functools import partial
+
 import numpy as np
 
 from qiskit_dynamics.signals import Signal, DiscreteSignal, DiscreteSignalSum, SignalList
 from qiskit_dynamics.signals.signals import to_SignalSum
-from qiskit_dynamics.array import Array
+from qiskit_dynamics.arraylias import DYNAMICS_NUMPY as unp
 
-from ..common import QiskitDynamicsTestCase, TestJaxBase
+from ..common import JAXTestBase, QiskitDynamicsTestCase, test_array_backends
 
 try:
     from jax import jit, grad
@@ -30,11 +32,12 @@ except ImportError:
     pass
 
 
+@partial(test_array_backends, array_libraries=["numpy", "jax", "array_numpy", "array_jax"])
 class TestSignal(QiskitDynamicsTestCase):
     """Tests for Signal object."""
 
     def setUp(self):
-        self.signal1 = Signal(lambda t: 0.25, carrier_freq=0.3)
+        self.signal1 = Signal(lambda _: 0.25, carrier_freq=0.3)
         self.signal2 = Signal(lambda t: 2.0 * (t**2), carrier_freq=0.1)
         self.signal3 = Signal(lambda t: 2.0 * (t**2) + 1j * t, carrier_freq=0.1, phase=-0.1)
 
@@ -51,25 +54,27 @@ class TestSignal(QiskitDynamicsTestCase):
 
     def test_envelope_vectorized(self):
         """Test vectorized evaluation of envelope."""
-        t_vals = np.array([1.1, 1.23])
-        self.assertAllClose(self.signal1.envelope(t_vals), np.array([0.25, 0.25]))
+        t_vals = self.asarray([1.1, 1.23])
+        self.assertAllClose(self.signal1.envelope(t_vals), self.asarray([0.25, 0.25]))
         self.assertAllClose(
-            self.signal2.envelope(t_vals), np.array([2 * (1.1**2), 2 * (1.23**2)])
+            self.signal2.envelope(t_vals), self.asarray([2 * (1.1**2), 2 * (1.23**2)])
         )
         self.assertAllClose(
             self.signal3.envelope(t_vals),
-            np.array([2 * (1.1**2) + 1j * 1.1, 2 * (1.23**2) + 1j * 1.23]),
+            self.asarray([2 * (1.1**2) + 1j * 1.1, 2 * (1.23**2) + 1j * 1.23]),
         )
 
-        t_vals = np.array([[1.1, 1.23], [0.1, 0.24]])
-        self.assertAllClose(self.signal1.envelope(t_vals), np.array([[0.25, 0.25], [0.25, 0.25]]))
+        t_vals = self.asarray([[1.1, 1.23], [0.1, 0.24]])
+        self.assertAllClose(
+            self.signal1.envelope(t_vals), self.asarray([[0.25, 0.25], [0.25, 0.25]])
+        )
         self.assertAllClose(
             self.signal2.envelope(t_vals),
-            np.array([[2 * (1.1**2), 2 * (1.23**2)], [2 * (0.1**2), 2 * (0.24**2)]]),
+            self.asarray([[2 * (1.1**2), 2 * (1.23**2)], [2 * (0.1**2), 2 * (0.24**2)]]),
         )
         self.assertAllClose(
             self.signal3.envelope(t_vals),
-            np.array(
+            self.asarray(
                 [
                     [2 * (1.1**2) + 1j * 1.1, 2 * (1.23**2) + 1j * 1.23],
                     [2 * (0.1**2) + 1j * 0.1, 2 * (0.24**2) + 1j * 0.24],
@@ -104,10 +109,10 @@ class TestSignal(QiskitDynamicsTestCase):
 
[…]
