+++ qiskit_dynamics/signals/signals.py
@@ -2,7 +2,7 @@
 
 # This code is part of Qiskit.
 #
-# (C) Copyright IBM 2021.
+# (C) Copyright IBM 2023.
 #
 # This code is licensed under the Apache License, Version 2.0. You may
 # obtain a copy of this license in the LICENSE.txt file in the root directory
@@ -24,13 +24,10 @@ import operator
 import numpy as np
 from matplotlib import pyplot as plt
 
-try:
-    import jax.numpy as jnp
-except ImportError:
-    pass
-
 from qiskit import QiskitError
-from qiskit_dynamics.array import Array
+from qiskit_dynamics.arraylias import ArrayLike
+from qiskit_dynamics.arraylias import DYNAMICS_NUMPY_ALIAS as alias
+from qiskit_dynamics.arraylias import DYNAMICS_NUMPY as unp
 
 
 class Signal:
@@ -72,9 +69,9 @@ class Signal:
 
     def __init__(
         self,
-        envelope: Union[Callable, complex, float, int, Array],
-        carrier_freq: Union[float, List, Array] = 0.0,
-        phase: Union[float, List, Array] = 0.0,
+        envelope: Union[Callable, complex, float, int, ArrayLike],
+        carrier_freq: Union[float, List, ArrayLike] = 0.0,
+        phase: Union[float, List, ArrayLike] = 0.0,
         name: Optional[str] = None,
     ):
         """
@@ -91,10 +88,9 @@ class Signal:
         self._name = name
         self._is_constant = False
 
-        if isinstance(envelope, (complex, float, int)):
-            envelope = Array(complex(envelope))
-
-        if isinstance(envelope, Array):
+        if not callable(envelope):
+            # if not callable, we assume a constant
+            envelope = unp.asarray(envelope)
             # if envelope is constant and the carrier is zero, this is a constant signal
             try:
                 # try block is for catching JAX tracer errors
@@ -103,15 +99,9 @@ class Signal:
             except Exception:  # pylint: disable=broad-except
                 pass
 
-            if envelope.backend == "jax":
-                self._envelope = lambda t: envelope * jnp.ones_like(Array(t).data)
-            else:
-                self._envelope = lambda t: envelope * np.ones_like(t)
-        elif callable(envelope):
-            if Array.default_backend() == "jax":
-                self._envelope = lambda t: Array(envelope(t))
-            else:
-                self._envelope = envelope
+            self._envelope = lambda t: envelope * unp.ones_like(t)
+        else:
+            self._envelope = envelope
 
         # set carrier and phase
         self.carrier_freq = carrier_freq
@@ -128,45 +118,41 @@ class Signal:
         return self._is_constant
 
     @property
-    def carrier_freq(self) -> Array:
+    def carrier_freq(self) -> ArrayLike:
         """The carrier frequency of the signal."""
         return self._carrier_freq
 
     @carrier_freq.setter
-    def carrier_freq(self, carrier_freq: Union[float, list, Array]):
+    def carrier_freq(self, carrier_freq: Union[float, list, ArrayLike]):
         """Carrier frequency setter. List handling is to support subclasses storing a
         list of frequencies."""
-        if type(carrier_freq) == list:
-            carrier_freq = [Array(entry).data for entry in carrier_freq]
-        self._carrier_freq = Array(carrier_freq)
+        self._carrier_freq = unp.asarray(carrier_freq)
         self._carrier_arg = 1j * 2 * np.pi * self._carrier_freq
 
     @property
-    def phase(self) -> Array:
+    def phase(self) -> ArrayLike:
         """The phase of the signal."""
         return self._phase
 
     @phase.setter
-    def phase(self, phase: Union[float, list, Array]):
[…]
