### 2023-10-16 d0d52ae
moving real casting inside of sum in jit_grad_wrap test function
