+++ test/dynamics/common.py
@@ -285,5 +285,5 @@ class TestJaxBase(unittest.TestCase):
             JIT-compiled gradient of function.
         """
         wf = wrap(lambda f: jit(grad(f)), decorator=True)
-        f = lambda *args: np.sum(func_to_test(*args).real)
+        f = lambda *args: np.sum(func_to_test(*args)).real
         return wf(f)
