### 2023-10-17 8872374
update test_array_backends infrastructure
