### 2023-10-17 6c9524c
updating function doc string
