### 2023-10-18 c595d8d
changing RoughFrequencyCal qubit argument to a list (#271)
