### 2023-10-19 bb5422c
pylint: disable=no-member in test_signals.py
