### 2023-10-19 47a9de4
added docstrings to setup in test_signals.py
