### 2023-10-20 9634cf5
adding test case that should fail when perturbation workaround no longer required
