### 2023-10-20 77e5142
adding release note for jax update
