+++ releasenotes/notes/update-jax-a50ce1b7d6b47219.yaml
@@ -1,0 +1,10 @@
+---
+issues:
+  - |
+    A JAX warning about casting complex values to real is raised when when jit-compiling gradients
+    of simulations in Qiskit Dynamics. Note that this warning does not appear to signify any error
+    in numerical computation, and can be safely ignored.
+upgrade:
+  - |
+    The upper bound on JAX and Diffrax in the last version of Qiskit Dynamics has been removed.
+    Users should try to use the latest version of JAX.
