### 2023-10-20 0138824
correcting release note
