+++ releasenotes/notes/update-jax-a50ce1b7d6b47219.yaml
@@ -1,8 +1,8 @@
 ---
 issues:
   - |
-    A JAX warning about casting complex values to real is raised when when jit-compiling gradients
-    of simulations in Qiskit Dynamics. Note that this warning does not appear to signify any error
+    A JAX warning about casting complex values to real is raised when when computing gradients of
+    simulations in Qiskit Dynamics. Note that this warning does not appear to signify any error
     in numerical computation, and can be safely ignored.
 upgrade:
   - |
