### 2023-10-22 #5853 @hodgestar
I have a question about the list of authors, partly for my own edification: The authors list in the paper is much shorter than the list of [GitHub contributors](https://github.com/Qiskit-Extensions/qiskit-dynamics/graphs/contributors). Some of the omitted contributors seem to have made significant contributions. How should one decide who should be listed as an author in the paper?
