### 2023-10-22 Qiskit-Extensions/qiskit-dynamics#274 @hodgestar
Make relationship between signals and hamiltonian_operators explicit in tutorial.
In https://qiskit.org/ecosystem/dynamics/tutorials/Rabi_oscillations.html the relationship between the `hamiltonian_operators`, the `signals` passed to the solver and the time-dependent term in the Hamiltonian isn't explicitly stated. I think readers who are not familiar with how such solvers structure things might not easily pick this up.

I also find the name `hamiltonian_operators` for the time-dependent / signal operators a bit strange. The static part of the Hamiltonian is also a Hamiltonian operator. Perhaps naming the argument `signal_operators` might this clearer?
