### 2023-10-22 #5853 @hodgestar
Should the references for the related open source packages be added to the list of references at the end of the paper?
