### 2023-10-22 #5853 @hodgestar
@danielskatz No related to this review, but it would be great if the JOSS papers could include small usage examples and outputs so that people reading them could see the package in action if they wanted to.
