### 2023-10-22 #5853 @hodgestar
@danielskatz Done with reviewing for the moment. The package looks great and introduces some nice higher level tools for users (e.g. the perturbation theory calculations and the signal abstraction) and, of course, provides first class integration with Qiskit.

Just waiting for your feedback on my two "what is the process" questions about which contributors should be included in the authors list and what should happen with citations for other packages.
