### 2023-10-22 #5853 @hodgestar
Thank you for including me in the review process. It was fun to explore qiskit-dynamics as part of the review and to get to know the JOSS workflow.
