### 2023-10-23 #5853 @danielskatz
> @danielskatz No related to this review, but it would be great if the JOSS papers could include small usage examples and outputs so that people reading them could see the package in action if they wanted to.

Please feel free to add an issue about this to https://github.com/openjournals/joss/issues
