### 2023-10-23 #5853 @danielskatz
> I have a question about the list of authors, partly for my own edification: The authors list in the paper is much shorter than the list of [GitHub contributors](https://github.com/Qiskit-Extensions/qiskit-dynamics/graphs/contributors). Some of the omitted contributors seem to have made significant contributions. How should one decide who should be listed as an author in the paper?

This is up to you and the authors to discuss. When there's a question like this, typically a reviewer asks the authors, who reply, then they see if they agree.
