### 2023-10-23 #5853 @danielskatz
> Should the references for the related open source packages be added to the list of references at the end of the paper?

The references at the end are autogenerated from those cited in the paper, so I think this question is really if there are missing citations in the text. If you believe there are, please suggest adding them to the author.
