### 2023-10-23 #5853 @hodgestar
Regarding the missing references -- they were on the next page of references. Doh. Resolved.
