### 2023-10-23 Qiskit-Extensions/qiskit-dynamics#275 @hodgestar
JOSS authorship list
Re: https://github.com/openjournals/joss-reviews/issues/5853

Hi @DanPuzzuoli.

I have a question about the list of authors in the JOSS paper, partly for my own edification:

The authors list in the JOSS paper is much shorter than the list of [GitHub contributors](https://github.com/Qiskit-Extensions/qiskit-dynamics/graphs/contributors). Some of the omitted contributors seem to have made significant contributions. How did you decide who should be listed as an author in the paper? Should any of the other contributors be added?

I have no strong opinions about who should be listed as authors, but thought it was worthwhile double checking.

Regards,
Simon
