### 2023-10-23 #5853 @danielskatz
@DanPuzzuoli - I've suggested some minor changes in https://github.com/Qiskit-Extensions/qiskit-dynamics/pull/276  Please merge this, or let me know what you disagree with, then we can proceed to acceptance and publication.
