### 2023-10-23 b9d0ebd
minor changes for JOSS paper (#276)
* minor changes in paper

* minor change in bib
