+++ paper.bib
@@ -142,7 +142,7 @@
 
 @misc{aer,
     author = {{Qiskit contributors}},
-    title = {Qiskit Aer},
+    title = {Qiskit {A}er},
     year = {2023},
 	url = {https://github.com/Qiskit/qiskit-aer},
 }
