+++ paper.md
@@ -37,9 +37,9 @@ bibliography: paper.bib
 
 # Summary
 
-Qiskit Dynamics is an open source Python library for numerically simulating the time dynamics of finite-dimensional quantum systems. The goal of the package is to provide flexible configuration of the numerical methods used for simulation: general tools for transforming models of quantum systems for more efficient simulation (rotating frames and the rotating wave approximation), choice of array representations (dense vs. sparse, and different array libraries), and access to different types of underlying solvers (standard ODE vs. geometric solvers). The package also contains advanced functionality for computing time-dependent perturbation theory expressions used in robust quantum control optimization [@perturb1; @perturb2].
+Qiskit Dynamics is an open-source Python library for numerically simulating the time dynamics of finite-dimensional quantum systems. The goal of the package is to provide flexible configuration of the numerical methods used for simulation: general tools for transforming models of quantum systems for more efficient simulation (rotating frames and the rotating wave approximation), choice of array representations (dense vs. sparse, and different array libraries), and access to different types of underlying solvers (standard ODE vs. geometric solvers). The package also contains advanced functionality for computing time-dependent perturbation theory expressions used in robust quantum control optimization [@perturb1; @perturb2].
 
-As part of the Qiskit Ecosystem (https://qiskit.org/ecosystem), the package interfaces with other parts of Qiskit [@Qiskit]. Most notably, Qiskit Dynamics provides tools for simulating control sequences specified by Qiskit Pulse [@alexander_qiskit_2020] which gives a pulse representation of quantum circuit instructions. Higher level interfaces allow users to build and interact with simulation-based objects that target the same constraints (qubit layout, control sequence sampling rate, etc.) as a specified IBM Quantum computer.
+As part of the Qiskit Ecosystem (https://qiskit.org/ecosystem), the package interfaces with other parts of Qiskit [@Qiskit]. Most notably, Qiskit Dynamics provides tools for simulating control sequences specified by Qiskit Pulse [@alexander_qiskit_2020], which gives a pulse representation of quantum circuit instructions. Higher-level interfaces allow users to build and interact with simulation-based objects that target the same constraints (qubit layout, control sequence sampling rate, etc.) as a specified IBM Quantum computer.
 
 Lastly, to facilitate high-perfomance applications, Qiskit Dynamics is compatible with the JAX array library [@jax2018github]. As such, all core computations are just-in-time compilable, automatically differentiable, and executable on GPU.
 
@@ -51,14 +51,14 @@ Furthermore, having a simulation-based drop-in replacement for real quantum comp
 
 # Related open source packages
 
-Due to its importance, many open source packages contain time-dependent quantum system simulation tools. In Python, these include QuTiP [@qutip], TorchQuantum [@torchquantum], and C3 [@C3]. C++ packages (also with Python interfaces) include lindbladmpo [@lindbladmpo] and Quandary [@quandary]. Packages also exist in other languages, such as the Hamiltonian open quantum system toolkit (HOQST) [@hoqst] and a Framework for Quantum Optimal Control [@julia_qc] in Julia, and Spinach [@spinach] in MATLAB. The features in Qiskit Dynamics for simulating Qiskit Pulse control sequences replace those previously offered in Qiskit Aer [@aer].
[…]
