### 2023-10-23 Qiskit-Extensions/qiskit-dynamics#274 @DanPuzzuoli
I agree the naming is a little odd. It's been a while so I can't remember the alternatives we had considered, but I think we settled on the current naming conventions as a compromise between clarity and wanting to avoid the argument names being too long. (E.g. `signal_operators` isn't quite correct either as there are signals that can be supplied as coefficients for dissipator terms as well.)

At the stage I'm resistant to changing them as they've been with the package for so long. However, I think giving more explanation in the tutorial would be good. I've also just reread the `Solver` class documentation and that can be clarified as well.
