### 2023-10-23 #5853 @DanPuzzuoli
@goerz @hodgestar @babreu-ncsa  thank you for your reviews, and thank you @danielskatz for editing!

@danielskatz I plan to make a minor change to documentation based on a suggestion by @hodgestar [here](https://github.com/Qiskit-Extensions/qiskit-dynamics/issues/274) that was not a blocker on their review. Aside from this, there will have been no changes made based on the reviews. Is it okay if I just stick with the latest tagged version for the archiving (0.4.2, which is what the reviewers used for review) and have this change included in the next natural release? I wouldn't normally make a new release solely for a change like this, and it'd be nice to conclude the process for the paper.
