### 2023-10-23 #5853 @DanPuzzuoli
@danielskatz

- The version tag of the reviewed software is [0.4.2](https://github.com/Qiskit-Extensions/qiskit-dynamics/releases/tag/0.4.2)
- DOI for the zenodo archive of the above version: [10.5281/zenodo.10034642](https://zenodo.org/doi/10.5281/zenodo.10034642) - I've edited the title and author list to match the paper.
