### 2023-10-24 #5853 @danielskatz
Congratulations to @DanPuzzuoli (Daniel Puzzuoli) and co-authors on your publication!!

And thanks to @babreu-ncsa, @goerz, and @hodgestar for reviewing!
JOSS depends on volunteers and couldn't be successful without you
